+++ sistemaGestionBiblioteca/src/main/java/app/biblioteca/modelos/Libro.java
@@ -1,6 +1,4 @@
-package app.biblioteca.recursos;
-
-import app.biblioteca.utils.Estado;
+package app.biblioteca.modelos;
 
 public class Libro {
     private String isbn;
